+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Support for DELETE /2.0/reports/{ReportId}/scope endpoint
 - Added wiremock integration tests for POST /2.0/reports/{reportId}/scope endpoint
 - Added wiremock integration tests for DELETE /2.0/reports/{reportId}/scope endpoint
+- Support for DELETE /2.0/reports/{reportId} endpoint
 
 ## [3.7.2] - 2026-01-29